+++ R/nuts_aggregate.R
@@ -6,7 +6,7 @@
 #' @param to_level Number corresponding to the desired NUTS level to be aggregated to: `1` or `2`.
 #' @param variables Named character specifying variable names and variable type (`'absolute'` or `'relative'`), e.g. `c('var_name' = 'absolute')`.
 #' @param weight String with name of the weight used for conversion. Can be area size `'areaKm'` (default),
-#' population in 2011 `'pop11'` or 2018 `'pop18'`, or artificial surfaces in 2012 `'artif_surf12'` and 2018 `'artif_surf18'`.
+#' population in 2011 `'pop11'` or 2021 `'pop21'`, artificial surfaces in 2012 `'artif_surf12'` and 2018 `'artif_surf18'` or residential built-up volume 2021 `'bu_vol'`.
 #' @param missing_rm Boolean that is FALSE by default. TRUE removes regional flows that depart from missing NUTS codes.
 #' @param missing_weights_pct Boolean that is FALSE by default. TRUE computes the percentage of missing weights due to missing departing NUTS regions for each variable.
 #' @param multiple_versions By default equal to `'error'`, when providing multiple NUTS versions within groups.
@@ -34,7 +34,7 @@
 #'   # Group vars are automatically passed on
 #'   nuts_aggregate(to_level = 2,
 #'                  variables = c('values'= 'absolute'),
-#'                  weight = 'pop18')
+#'                  weight = 'pop21')
 #'
 #' @export
 nuts_aggregate <- function(data,
@@ -74,10 +74,10 @@ nuts_aggregate <- function(data,
       cli_abort("Input {.arg to_level} invalid. Must be 1 or 2.")
 
     if (!is.null(weight)) {
-      if ((!any(weight %in% c("areaKm", "pop18", "pop11", "artif_surf18",
-                         "artif_surf12"))) | length(weight) >  1)
+      if ((!any(weight %in% c("areaKm", "pop21", "pop11", "artif_surf18",
+                         "artif_surf12", "bu_vol"))) | length(weight) >  1)
         cli_abort(
-          "Input {.arg weight} invalid. Must be either 'areaKm', 'pop11', 'pop18', 'artif_surf12' or 'artif_surf18'."
+          "Input {.arg weight} invalid. Must be either 'areaKm', 'pop11', 'pop21', 'artif_surf12', 'artif_surf18' or 'bu_vol'."
         )
     }